Create ZHcc_Correlated_Efficiencies_Plots
These plots build on the original efficiency plot script, but now consider the correlated efficiency values for b- and l-tagging efficiency that we get from the 2D histogram in "Search for the Decay of the Higgs Boson to Charm Quarks with the ATLAS Experiment" (The ATLAS Collaboration, Phys. Rev. Lett. DOI:10.1103/PhysRevLett.120.211802, 2018.)